requirements.txt
requests
python-dotenv
tqdm
pydantic